Practical work 3: Redesign Mark feature
This commit modifies the architecture to focus on mark management. In
general, Student and Course, which are working with marks in this
application, become the subclass of MarkManager, so both of them can
manage the mark in their own ways. As such, other affected
attributes/features are also rewritten to adapt this change.